requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.1
Django==2.2.7
django-appconf==1.0.3
django-crispy-forms==1.8.0
django-filter==2.2.0
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
jsonfield==3.1.0
kombu==4.6.8
Pillow==6.2.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
swapper==1.1.2.post1
vine==1.3.0